requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
click==8.1.7
    # via typer
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pillow==10.3.0
pygments==2.18.0
    # via rich
pyyaml==6.0.1
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
typer==0.12.3
typing-extensions==4.11.0
    # via typer
